final hehe

+++ src/main/kotlin/com/crm/corecrm/domain/model/Chat.kt
@@ -1,9 +1,9 @@
 package com.crm.corecrm.domain.model
 
 data class Chat(
-    val id: Long? = null,
-    val tgChatId: Long,
-    val creatorBy: Long,
-    val createdAt: Long,
+    val id: Int? = null,
+    val tgChatId: Int,
+    val creatorBy: Int,
+    val createdAt: Int,
     val status: ChatStatus,
 )
